requirements.txt
firebase_admin
Flask
Flask-Assets
Flask-Classful
google-cloud-storage
gunicorn
jsmin
matplotlib
mistune
pygal
pylint
pytest
python-dotenv
quandl
requests
simplejson
yapf
Flask-GoogleMaps